Backend/pyproject.toml
[project]
name = "speaksmart-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "boto3>=1.42.89",
    "fastapi>=0.135.2",
    "firebase-admin>=7.3.0",
    "google-cloud-texttospeech>=2.36.0",
    "httpx>=0.28.1",
    "librosa>=0.11.0",
    "numpy>=2.4.4",
    "openai-whisper>=20250625",
    "passlib[bcrypt]>=1.7.4",
    "pykakasi>=2.3.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "scipy>=1.17.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "torch>=2.11.0",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
